Support for datapath assisted mac address learning
If mac learning is enabled, whenever vRouter receives a packet, it
will check whether the source mac is already present in its bridge
table. If there is no entry for the mac address it will create an
entry in the bridge table and send a trap to the agent to validate
the entry and fix its nexthop.

In order to make sure that vRouter does not send multiple trap
messages to the agent when packets carrying the same source mac lands
up in different cpus, vRouter calclulates a hash of the source mac and
uses it to index to a particular cpu and acquires a bit lock for that
cpu in the ingress interface to prevent parallel learning/creation of
mac entries.

Also, if a 'mac-move' happens, vRouter will detect such a move and
trap the packet to the agent to get an updated view.

MAC address learning for packets coming from fabric interface needs
to be ironed out.

Partial-Bug: #1636413

+++ dpdk/vr_dpdk_host.c
@@ -487,7 +487,7 @@ dpdk_get_time(unsigned long *sec, unsigned long *usec)
 }
 
 static void
-dpdk_get_mono_time(unsigned int *sec, unsigned int *nsec)
+dpdk_get_mono_time(unsigned long *sec, unsigned long *nsec)
 {
     struct timespec ts;
